Revert "feat: 사장님용 직원 관리 화면 구현"
