+++ app/build.gradle.kts
@@ -49,7 +49,6 @@ dependencies {
     implementation(projects.feature.mypage)
     implementation(projects.feature.bosshome)
     implementation(projects.feature.bossmypage)
-    implementation(projects.feature.bossworkermanage)
 
     // cores
     implementation(projects.core.model)
