+++ core/data/src/main/java/com/uuranus/data/di/DataModule.kt
@@ -8,14 +8,10 @@ import com.uuranus.data.repository.mypage.MyPageRepository
 import com.uuranus.data.repository.mypage.MyPageRepositoryImpl
 import com.uuranus.data.repository.user.UserRepository
 import com.uuranus.data.repository.user.UserRepositoryImpl
-import com.uuranus.data.repository.workermanage.WorkerManageRepository
-import com.uuranus.data.repository.workermanage.WorkerManageRepositoryImpl
 import com.uuranus.myschedule.core.network.datasource.MyPageDataSource
 import com.uuranus.myschedule.core.network.datasource.MyPageDataSourceImpl
 import com.uuranus.myschedule.core.network.datasource.ScheduleDataSource
 import com.uuranus.myschedule.core.network.datasource.ScheduleDataSourceImpl
-import com.uuranus.myschedule.core.network.datasource.WorkerDataSource
-import com.uuranus.myschedule.core.network.datasource.WorkerDataSourceImpl
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -26,24 +22,17 @@ import dagger.hilt.components.SingletonComponent
 internal interface DataModule {
 
     @Binds
-    fun bindScheduleDataSource(
+    fun bindCalendarDataSource(
         dataSource: ScheduleDataSourceImpl,
     ): ScheduleDataSource
 
     @Binds
-    fun bindWorkerDataSource(
-        dataSource: WorkerDataSourceImpl,
-    ): WorkerDataSource
-  
     fun bindMyPageDataSource(
         dataSource: MyPageDataSourceImpl,
     ): MyPageDataSource
 
     @Binds
     fun bindsHomeRepository(homeRepositoryImpl: HomeRepositoryImpl): HomeRepository
-
-    @Binds
-    fun bindsWorkerRepository(workerRepositoryImpl: WorkerManageRepositoryImpl): WorkerManageRepository
 
     @Binds
     fun bindsLoginRepository(loginRepositoryImpl: LoginRepositoryImpl): LoginRepository
