+++ core/data/src/main/java/com/uuranus/data/repository/home/HomeRepository.kt
@@ -33,6 +33,10 @@ interface HomeRepository {
         storeMemberAvailableTimeId: Int,
     ): String
 
+    suspend fun getAllWorkers(
+        storeId: Int,
+    ): List<WorkerInfo>
+
     suspend fun changeSchedule(
         storeId: Int,
         scheduleUpdate: ScheduleUpdate,
