+++ core/domain/src/main/java/com/uuranus/domain/GetAllWorkersInfo.kt
@@ -1,13 +1,12 @@
 package com.uuranus.domain
 
 import com.uuranus.data.repository.home.HomeRepository
-import com.uuranus.data.repository.workermanage.WorkerManageRepository
 import com.uuranus.model.MyPossibleTimeInfo
 import com.uuranus.model.WorkerInfo
 import javax.inject.Inject
 
 class GetAllWorkersInfo @Inject constructor(
-    private val repository: WorkerManageRepository,
+    private val repository: HomeRepository,
 ) {
     suspend operator fun invoke(
         storeId: Int,
