+++ core/network/src/main/java/com/uuranus/myschedule/core/network/datasource/ScheduleDataSource.kt
@@ -34,6 +34,10 @@ interface ScheduleDataSource {
         storeMemberAvailableTimeId: Int,
     ): String
 
+    suspend fun getAllWorkers(
+        storeId: Int,
+    ): List<WorkerInfo>
+
     suspend fun changedSchedule(
         storeId: Int,
         scheduleUpdate: ScheduleUpdate,
