+++ settings.gradle.kts
@@ -34,4 +34,3 @@ include(":core:network")
 include(":feature:bosshome")
 include(":feature:bossmypage")
 include(":core:common")
-include(":feature:bossworkermanage")
